pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dante-Lucas <lucassilva270805@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
django = "^5.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.13.0"
ruff = "^0.6.8"

[tool.taskipy.tasks]
lint = 'ruff check .'
diff = 'ruff check . --diff'
fix = 'ruff check . --fix'
format = 'ruff format .'

server = "python manage.py runserver"
make = "python manage.py makemigrations"
migrate = "python manage.py migrate"
create-admin = "python manage.py createsuperuser"
static = "python manage.py collectstatic"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
